circe-command-TOPIC: fix bug when called with no arg
Previously this command would cause an irc error when it was called with
no arg.  That is because it was incorrectly testing for a nil value of
'channel' when it should have been testing for the empty string.

I considered modifying 'circe-chat-input' instead to supply commands with
nil when called at the circe prompt with no argument, but subsequently
rationalized that the empty string is the correct thing to pass because,
if IRC protocol is considered to be a minilanguage, the empty string is
its null value.  A number of existing commands are already built to
exploit this feature, and would need to be rewritten in a more complex
fashion if nil were used instead of the empty string for "no argument".