circe-lagmon.el: a new lag monitor mode
Start the lag monitor by loading this module and calling
circe-lagmon-mode.  It monitors your lag on each server by periodically
sending a CTCP LAGMON to your own nick, bearing a timestamp.  It also
monitors for timeouts of these messages, and can try to automatically
reconnect when one times out.

Configuration is controlled by the following variables:

  circe-lagmon-interval: sets the main interval for lag monitoring.
Default is 60 (seconds).

  circe-lagmon-mode-line-format-string: format string for the mode-line
display.

  circe-lagmon-reconnect-timeout: seconds after which to try reconnecting
to the server when a CTCP LAGMON message times out.  Default is 20.