shorten JOIN and QUIT messages
Removed redundant text from the JOIN and QUIT messages.  Since the
messages begin with "*** Join:" and "*** Quit:", it is redundant and a
waste of space to say at the end "is now on the channel" or "has left
IRC".

Also update format of PART message to be consistent with that of QUIT.