circe-display: circe-server-face and circe-my-message-face low precedence
This patch makes circe-server-face and circe-my-message-face have the
lowest precedence in the faces stack for messages.  This means that when
another face is present, such as circe-originator-face or a face applied
via *circe-special-display-spec*, those faces take precedence.

Special-display faces have low precedence (just above circe-server-face
and circe-my-message-face), except when the 'highlight-area property is
used to limit the area that the special-display face is applied to; then
the special-display face has higher precedence than faces that came from
keywords (like circe-originator-face).  The rationale is that if
'highlight-area is present, special-display is being used to call
attention to a specific part of the message, so, higher specificity means
higher precedence.