circe-format-self-action: use 'nick' instead of 'mynick'
circe-command-ME: supply ':nick' keyword

The rationale for this change is that when lui-format propertizes
circe-format-self-action with the lui-format-argument, it makes more sense
semantically for that object to be marked as a 'nick' than as a 'mynick'.
Nick colorization is an application that may make use of these properties,
and whose implementation is simplified by this change.