auto-join: Implement new behavior using only a single channel list.
We had developed three different auto-join lists by now, which is
a bit heavy. This unifies the different types into a single list
again and is extensible for the future in case we find other
situations in which we want to auto-join, too. It also has a
single point for joining in case we need to customize auto-joining
further.

Thanks to mtp for coming up with the idea for this patch.