fools list: Also handle ACTION messages and ignore properly (fixes #4)
The ACTION handler is bypassing the default server display (this
smells a bit like bad design), and thus didn't handle fools
properly. Additionally, the tracking code in lui didn't check
if *any* point in the new message had a 'lui-fool property. I'm
not sure if that really was the problem with fool messages showing
up in the tracking buffer, but it sounds possible. This is now
both fixed.