Completion: Switch to `completion-at-point' (fixes #3)
Emacs 23 introduced a new completion framework using a single
list of functions to generate completions. This obsoletes the
plethora of completion libraries in use before, including
lui's incomplete.el library. This patch switches lui to use
the new interface. The application interface of Lui is not
changed, so other programs will not have to change anything
to make use of this.