joinpart spam: Use `circe-channel-users' (closes #5)
With this patch, `circe-channel-users' now maps nicks to
a hash table. This second table can be used to store extra
information for a given nick. The first application for this
is the joinpart spam protection, which stores a lurker flag
that is nil if the user is no lurker (default should be t).

Circe by default also stores a joined time in the hash table,
the `float-time' when this user joined a channel, or when we
joined if the user has been there longer than us.