lagmon: Handle nick changes, and some minor fixes
If the user changes a nick just after we sent a CTCP LAGMON,
we might miss the response. Hence, we simply claim we never
sent a CTCP message after a nick change. This causes extra
CTCP messages after nick changes, but that shouldn't be a
problem.

We also increase the lag timeout to 40s, as the old value was
a bit short during connection bursts. This is equivalent to
20s of lag. That could be confusing, but I think the idea of
"disconnect after not receiving a reply for 40s" is a bit
easier, conceptually.

Also renamed circe-lagmon-mode-line-no-lag-format-string to
circe-lagmon-mode-line-unknown-lag-indicator to better reflect
the purpose.