circe-server-auto-join-channels: keyword in first position used for manually joined channels
This patch addresses the following scenario:

  You are on two networks, for one of which you have nickserv credentials,
  the other not.  On the network with nickserv, you want all of your
  autojoins to happen :after-auth, including rejoining of manually joined
  channels.  On the non-nickserv network, you want :immediate auto-join
  for all channels.

Previously, it was not possible to do the above, because for the sake of
the non-nickserv network, circe-server-auto-join-default-type would have
to be :immediate, and that setting would be used for manually joined
channels on the with-nickserv network.  This patch adds the following
convention to address this situation:

A type keyword in the first position of a channel list in
circe-server-auto-join-channels is used for manually joined channels.
Thus the situation above can be addressed by leaving
circe-server-auto-join-default-type at it's default value of :immediate,
and putting :after-auth in the first position of the auto-join channels
list of the with-nickserv network.