circe-reconnect: Make `open-tls-stream` use a temporary buffer.
As `open-tls-stream` is spamming its associated buffer with status
messages, we let it do so in a temporary buffer and associate the
process with the server buffer only after we set up a process
filter function.

This is a rework of the fix done in 5b7d642, and hopefully a lot
cleaner.