lui-scroll-to-bottom: Always keep the input line at the bottom.
The old code allowed the window to scroll past the end of the buffer
if point was not in the input line. That was behavior copied from
ERC. I'm unsure why that was done in the first place. The new code
does not differentiate between point being in the input line or in the
rest of the IRC buffer.