lui-autopaste: New Lui extension to use a paste service for long input.
The lui-autopaste extension will intercept long input to lui and
ask the user if they would prefer to use a paste service instead.
If they agree, the input is replaced with the URL to the paste
contents.

By default, bpaste.net is used, but other paste services can be
hooked into the extension.