Revert "Flyspell: Mark current channel's nicks as correctly spelled (closes #9)."
This is not trivial to implement, because flyspell does not recognize
all nicks as words. The solution employed here is suboptimal to
begin with and can lead to problems when nicks use numbers. A better
solution would be for Circe to check itself if point is at a nick,
but that's non-trivial as well, especially as flyspell decides how
and when to call that hook.

This bug requires a lot more thinking.

This reverts commit 5d3b513538cf37a37f6bb9a7466633d90861feb6.