lui-scroll: Support different scrolling types to accomodate Emacs.
Since Emacs 24, the old scroll-to-bottom behavior using
`window-scroll-functions' is recommended against directly in the
docstring of that variable. There is at this time no good solution
to achieve this. Various Emacs modes implement different solutions.
This commit makes Circe use any of the popular ones, defaulting
to the most compliant one.

The compliant behavior, setting `lui-scroll-behavior` to `post-output`,
should also solve speed problems with the old solution and therefore
fixes #25.