circe-reconnect: Use the sentinel for TLS errors, too.
TLS connections do not use the process sentinel on connect, so
we do not reconnect when an error occurs. This commit changes
the code of tls connection to call the sentinel on a connection
error, too. As a most important effect, this means TLS connections
also honor `circe-server-max-reconnect-attempts`, and thus
fixes #2.