circe-channel-receiving-names: Do not clear channel user information on NAMES.
When Circe received a NAMES message, we cleared all user information.
This happens usually on a JOIN, where it is sensible. When looking
at currently active users in a channel, this actually marked them
all as "just joined" and "non-lurking", which is a bad side effect
of just checking who's on the channel. This commit synchronizes
the channel names list with the internal list on an additional
NAMES reply, but does not clear information on users we already
have.

Fixes #32.