circe-display: 'highlight-area property to limit special-display face
circe-special-display-alist may cause a face to be added to a message
based on its sender.  This patch allows the portion of the message that
the face is applied to to be limited to characters with the text property,
'highlight-area t.  A format function (instead of a format string) can be
used to add this property.