circe-auto-join: fix bug related to circe-new-buffer-behavior-ignore-auto-joins
The circe-new-buffer-behavior-ignore-auto-joins mechanism works by keeping
a list of auto-joined channels, and by circe-server-get-chat-buffer
ignoring channels in that list in the cond branch that creates a channel
buffer if it does not exist.  This same branch removes the channels from
the list after they are used, so that a later JOIN of an auto-joined
channel, for example, if the user killed the channel buffer, then used the
JOIN command, would let circe-new-buffer-behavior be used to determine
whether and how the new buffer is displayed.

There was a bug, because if you reconnect to a network, channel buffers
already exist for all of the auto-join channels, so the cond branch that
manages the aforementioned list would never be reached.  The channel name
would not get cleaned out of the list when the channel was auto-joined,
and a subsequent join of that channel would not behave according to
circe-new-buffer-behavior.

This patch fixes that bug by only adding auto-joined channel names to the
list if a channel buffer for the channel does not already exist.