circe-default-properties: New name for circe-special-display.
This setting will set the default properties of a Circe display
message. The new name is a lot more intuitive. "Special" is always
a bit iffy to use.

The setting of *circe-default-properties* for CTCP events is also
moved from the CTCP ACTION handler to the general CTCP display
handler, where it belongs.