circe-channel-users-synchronize: Fix and add proper initial-name handling.
If we just mark the initial users as active, they will emit part/quit
messages even if they never talked.  But we also don't want Circe to
emit a "first activity" message for these users, so we give them a
property `initial-names' and check for that in the mark-as-active
function to omit their first-activity message.