circe-fool-face: not bold in X
In situations where this face is together with other faces in a single
message, text color may come from another face (depending on precedence),
but the bold attribute contributed by circe-fool-face then serves to
emphasize the message, counter to the intent of the 'fool' system.  This
change only affects X; the bold attribute is kept for tty, where styling
choices are more limited.