circe-lurker-p: Use the last-active information instead of lurker-p.
Storing 'lurker-p as a user info was quite contrieved, as it was
only ever set with value nil and then queried with a default value
of t to check for lurkerness. With the introduction of 'last-active,
we have a much better metric for that. This simplifies the code in
`circe-lurker-mark-as-active` as well.