Recent channel users: Rework the way we keep recent user info.
This commit changes circe-channel-recent-users to store copies of
the user hashes from circe-channel-users while those users are not
present. On re-join, the user info is copied back, just with a few
fields overridden.

Additionally, this is now correctly split up. The management of
those two hashes happens in `circe-channel-message-handler`, while
the decision on what to display or not happens in the display
handlers.

Related changes:

- circe-parted-users-timeout renamed to circe-channel-recent-users-timeout
- circe-format-channel-rejoin now has a user@host mask (fixes #47)