circe-display-NICK: Display a message on nick regain.
When a lurker joins and then changes their nick to one which had
recently left the channel, the nick change should be displayed, as the
new user will inherit the settings of the old nick.

This is not 100% accurate, as a user can quit and then another
occupies the nick. But as long as we don't attach authorization
information to the nick info, this should be an sufficient heuristic.