circe-reconnect: Clean up process.
When using openssl/gnutls as the connection provider, it was
possible for Circe to establish two connections to the same server
in the same server buffer, causing to a lot of very weird problems.

This was due to the process getting killed while the sentinel was
still active, causing another circe-reconnect to be run and only
then the current circe-reconnect to be finished.

We now clean up the server process by first removing the sentinel.

Fixes #66