Overhaul how nick colors are picked.
Instead of picking random numbers and seeing if they're appropriate,
precompute a pool of colors that are appropriate and also distinct from
each other.  The parameters for appropriateness and sufficient
distinctness are customizable.

After the pool is exhausted, pick the color that least recently
*appeared on screen* whenever a new nick color is needed.

The pool is generated to contain colors supported by your display.  If
you switch between displays supporting different color ranges, you may
want to call `circe-nick-color-reset'.

Get tty colors via (tty-color-alist) in advance instead of letting Emacs
compute them from the `color-name-rgb-alist', because it uses the costly
`tty-color-approximate' function to do so.