Added non-blocking TLS-connections support.
The `circe-tls-make-stream' function tries to mimic the
`make-network-process' function in it's parameters, but to behave
closely to standard `open-tls-stream'.

Warning: tls-success is not correct by default in Emacs 24.3.1 with
openssl 1.0.1e or 1.0.1f, and it's important for this new function,
so it should be changed to "- Handshake was completed\|Verify return
code".