Fixed some of circe-tls issues
Added a check for process attached to a target buffer, so now it must not
assign the same buffer to different processes, even if they are running
simultaneously.

And removed a call to fail-thunk from the temporary process filter, so now
only it's sentinel calls the fail-thunk (and it does not open excessive
connections on error).