Make search case-sensitive
Before this change there was a minor annoyance involved into the
matching of nicks in messages.  If the nick in question had uppercase
letters, but could still be found with an case-insensitive search in a
message, it got highlighted with the invalid foreground color nil
because the nick with the different upcasing couldn't be found in the
nick-color mapping.  This issue manifested itself in numerous "Invalid
face attribute :foreground nil" error messages in the *Messages* buffer.