Ignore process output with a filter, not another buffer.
Some TLS programs (openssl, gnutls) will emit status information when
killed, which would spam the server buffer if we kill the process.
Use a no-op filter to ignore this, instead of redirecting the output
to a temp buffer, as the latter can result in spurious errors with
marks in the wrong buffer.

Fixes #112