make-tls-process now returns a process argument.
Trying multiple commands in turn should not be necessary if we
figure out whether to use gnutls or openssl from the start. The only
additional thing the original tls.el code does it to try using SSLv3,
which we do not want to do anyhow.

This is to facilitate solving #130.