Ensure circe-server-buffer is set in chat mode hooks.
Channel and query modes should be customizeable through their mode
hooks, which requires a valid value of circe-server-buffer. But as
modes defined with define-derived-mode can not take arguments, they
can not take the server buffer as an argumen. As such modes also
kill all local variables as the first thing, it can't be passed that
way, either. So we pass them using dynamically-scoped variables
and use that as a protocol. Not pretty, but it works.

Fixes #165