Increase lagmons timeout delay.
Lagmon used to reconnect if it did not get a response within 40
seconds of sending a query. That's awfully short in some situations,
while pretty long in others. Generally, if we can still send data
to an IRC socket, we do not need to reconnect right away, as the
connection is still active. Worst-case, we reconnect to a heavily
clogged server, which is kind of rude. This commit bumps the timeout
to two minutes, which should be enough for detecting seriously
broken connections, but not fire too early in almost all cases.