Reconcile server buffers and display-buffer-alist
First of all, `switch-to-buffer` cannot be controlled by customizing
`display-buffer-alist`, so it's been replaced by the functionally
equivalent `pop-to-buffer-same-window` function.  The other thing to fix
was to ensure that the server buffer is no longer the selected buffer to
allow for examining buffer-local values of the buffer `circe` was
started from before displaying the server buffer.