Optimize nick listing function
The naïve rewrite of `circe-nick-color-nick-list` did just replace the
existing lookups with the equivalent functions provided by the irc.el
rewrite.  What wasn't taken into account is that both
`circe-server-my-nick-p` and `circe-channel-user-p` always calculate the
full list, then check.  As the checks happen repeatedly, the respective
lists can be calculated once instead and the lookups be made
significantly faster.  Due to this function being used for each message,
this change should make a difference for ZNC playback as this does bulk
inserts.