Filter dead buffers before constructing indicator
It can happen that a tracked buffer is killed, yet still lingers in
`tracking-buffers` and will be displayed as an empty item in the
tracking indicator.  While a window configuration change will have the
side effect of cleaning up that list (see
`tracking-remove-visible-buffers` for details), it's cleaner to filter
it when creating the status indicator.