Rework the lui button handling.
lui-buttons-list now can include URLs directly to avoid having to
write the same kind of function all the time. It can now also pass
multiple arguments to the function if needed.

Make use of this new functionality.

Also adjust the highlighting of issue numbers so there is no button
if the tracking URL is not configured.