Add edit and delete functions to lui
Protocols other than IRC might support editing and deleting messages,
such as the Slack chat application.

To facilitate, add an id to each incoming message, this is a
buffer-local incrementing number, and store the formatted and previous
time stamps in text properties so they can be restored later. Also store
each text property that gets stored using
‘lui-insert-with-text-properties’ so that when a message is edited the
right values are restored for those text properties.

The api is as follows:

- lui-replace-message: Replace the message at point with a given
  message.

- lui-replace: Replace a message with a given message. This variant
  takes a predicate function of no arguments that returns a non-nil
  value for the message to be replaced.

- lui-delete-message: Delete the message at point. Deleted messages are
  not removed from the buffer, but given a special face to indicate that
  they’ve been deleted. This helps visualize the deletion instead of
  suddenly disappearing from the buffer like it’s never been there.

- lui-delete: Delete a message. This variant takes a predicate function
  of no arguments that returns a non-nil value for the message to be
  replaced.

- lui-backward-message: Move backward one message and return point.

- lui-forward-message: Move forward one message and return point.