Improve track bar appearance
Currently the display of track bar is sub-optimal -- it is fixed width
and does not work well with time-stamps.

Instead of using a set width of space with strike-through face, use
background color on a newline -- this way Emacs fills the whole buffer
width with the face's background color.  The height of the line can be
customized by setting the face height value to a fraction.  This trick
also works on TTY.