Fix margin time stamp position
Put the time stamp display text at the beginning of the line when
‘lui-time-stamp-position’ is either left-margin or right-margin and
‘lui-fill-type’ is nil.

When ‘lui-fill-type’ has a value of nil, ‘visual-line-mode’ is enabled
and the ‘lui-time-stamp-position’ is either left-margin or
right-margin the time stamp appears in the margin of the last line of
the message. For all other situations it appears in the margin of the
first line of the message.