Alter buffer-undo-list manipulation to workaround emacs 25 GC change
In emacs 25 the GC seems to have changed: The old pattern used by
lui-adjust-undo-list was to mapcar over the undo list (returning
a new list which shared some elements with the old undo-list) and
assign the resulting list directly back into buffer-undo-list.

This seems to provoke an intermittent GC problem where strings in
the buffer-undo-list get double-freed, which makes emacs crash.

The implementation provided here uses the approach that erc uses
(which inspired the original lui-adjust-undo-list anyway) of
iterating over the items in buffer-undo-list and altering entries
where necessary using setf, incf and the like.