Shorten only the buffer names we have ever tracked
shorten-strings seems like it could be quadratic or worse, so having
fewer things to shorten could make things more responsive.

This is a win for users who have lots and lots of buffers, but
relatively few Circe buffers. It keeps track of all buffers that have
ever been tracked forever, so it could potentially be a loss for
people who keep very few buffers open, but cycle through lots of Circe
buffers.

Refs #286.