Shorten only lui-mode buffer names
shorten-strings seems like it could be quadratic or worse, so having
fewer things to shorten could make things more responsive.

This seems like a violation of the software's layered design (lui
knows about tracking, but tracking doesn't know about lui), but it's
just a symbol name so maybe it's OK. Another alternative is to maybe
define tracking-all-buffers as a hook in tracking and have lui provide
its definition. This makes the interface between lui and tracking a
bit more complicated, though.

Refs #286 and #344.