Use buffer-list-update-hook instead of advicing select-window
As requested for in a comment for the advice, since there seems to be
a hook for this.

The natural implementation of the hook would use
`with-circe-server-buffer`, but i think that it since
uses (indirectly, via `with-current-buffer`) `save-current-buffer`, it
alters the list of buffers ordering, something not expected by some
other functions adding hooks here.  In my case, it wreaks havoc with
exwm's input focus management, which uses this hook (advising
select-window has the same negative effect).