Allow customizing faces for queries and non-highlighted channels
At this point, we have a possibility to customize the color of a channel (in the
tracking segment of a modeline) where somebody mentioned our name. We use the
same face as for printing our name in the message itself
(`circe-highlight-nick-face`).

This is IMHO not sufficient because new personal messages are as important as
mentions in a channel and they can be easily missed when shown in the
default (for me gray) color. I am adding a support for this.

While I am at it, I am adding a possibility to customize a color of a channel,
that doesn't mention our name but has some new activity in it.

I understand that `tracking-add-buffer` allows adding buffers to
`tracking-buffers` with face and we may utilize this feature. I believe it makes
sense for what whatever it is currently used but I would prefer to have a
possibility to apply faces when rendering (in opposite to assigning a face when
some activity happens), hence `tracking-get-face`.

By default, I am setting the `circe-tracking-channel-face` and
`circe-tracking-query-face` to `nil` and therefore they are not going
to be customized and a backward-compatibility is going to be kept for
everybody who doesn't care about this feature. Personally, I am
putting the following lines to my config.

    (setq tracking-get-face-function #'circe-tracking-get-face)
    (set-face-attribute 'circe-tracking-channel-face nil :foreground my/white)
    (set-face-attribute 'circe-tracking-query-face nil :foreground my/blue)