Improve the implementation of SASL authentication
This PR adds support for all the possible events sent during SASL
authentication:

 - RPL_SASLSUCCESS
 - ERR_SASLFAILED
 - ERR_SASLTOOLONG
 - ERR_SASLABORTED
 - ERR_SASLALREADY

It also improves the registration logic in this way:

1. When using SASL on successful authentication the server will send 900 and
903. Right now on 900 the `sasl.login` event is always emitted even if we are
doing SASL authentication. This PR waits for the 903 when SASL negotiation is
ongoing. If SASL is not active the `sasl.login` will be emitted when 900 is
sent.

2. Right now the `CAP END` command is sent right after the `AUTHENTICATION` has
been sent without waiting for a response from the server. This PR also delays
the `CAP END` until we know the status of the SASL authentication (success,
failed or aborted).

3. It also adds a new flag `:sasl-strict` that makes the connection with the
server to be closed in the SASL authentication failed.