Add a logging network option
We currently have two ways to enable logging:
 - globally, with `(enable-lui-logging-globally)`
 - per buffer, with `M-x enable-lui-logging`

I believe we can make that a little easier and have an option
in `circe-network-option` to enable logging for a server:
```elisp
(setq circe-network-options '(
  ("Libera Chat"
   ;; ...
   :logging t)))
```

Most of the time I want to enable or disable logging for a server.
I guess if people want to have more granular options, we can easily
support a list of channels there too.

If this gets merged, I'll update
https://github.com/jorgenschaefer/circe/wiki/Configuration#logging