Implement spoiler faces
Previously: Using the same IRC color code on foreground/background
resulted in similar, but distinct foreground/background colors

Now: The same foreground/background color is used. To show the
spoiler, the text needs to be selected by activating the mark and
moving point to the other end. A future improvement (for GUI Emacs
specifically) would be to use 'mouse-face to display the text with the
same background/different foreground on hover, additionally
'cursor-face could be used in combination with
`cursor-face-highlight-mode` (to support textual Emacs, too).

Note that this does even work with overlapping patterns. For example
`<CODE><COLOR>visible<CODE><COLOR>,<COLOR>hidden<CODE><COLOR>visible<CODE>`
is displayed correctly.